front/requirements.txt
psycopg2-binary==2.9.9
Flask==3.0.0
pymongo==3.11.4
Pillow==10.1.0
requests==2.25.1
